llama-api-docker/RAG/requirements_light.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
sentence-transformers>=2.2.2
numpy>=1.20.0
psutil>=5.9.0
python-dotenv>=0.20.0 